final refactorings and README update
